+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Link to the cart on the Toaster success message
+
 ## [0.9.0] - 2020-06-12
 
 ### Added